Rename vuln parser package.
