+++ pkg/ingestor/parser/vuln/vuln_test.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package certify_vuln
+package vuln
 
 // TODO(bulldozer): freeze tests
 /*
